fix: update cashier redirection
